refactor(shared): consolidate modules and update biome schema path
