--- shared/constants.ts
@@ -1,12 +1,0 @@
-export const IMAGE_CONFIG = {
-	dimensions: {
-		padding: 40,
-	},
-	typography: {
-		headingMinSize: 32,
-		headingPercentage: 0.075, // 7.5% of max dimension
-		subheadingMinSize: 24,
-		subheadingPercentage: 0.055, // 5.5% of max dimension
-		lineSpacingMultiplier: 1.2,
-	},
-} as const;
