+++ shared/index.ts
@@ -1,5 +1,546 @@
-export * from "./analytics";
-export * from "./constants";
-export * from "./jobStatus";
-export * from "./metricPayload";
-export * from "./validators";
+// ===================================================================================
+// Image and Formatting Constants
+// ===================================================================================
+
+export const IMAGE_CONFIG = {
+	dimensions: {
+		padding: 40,
+	},
+	typography: {
+		headingMinSize: 32,
+		headingPercentage: 0.075, // 7.5% of max dimension
+		subheadingMinSize: 24,
+		subheadingPercentage: 0.055, // 5.5% of max dimension
+		lineSpacingMultiplier: 1.2,
+	},
+} as const;
+
+// ===================================================================================
+// Job Status Definitions
+// ===================================================================================
+
+export const JOB_STATUS_PENDING = "pending";
+export const JOB_STATUS_PROCESSING = "processing";
+export const JOB_STATUS_COMPLETED = "completed";
+export const JOB_STATUS_FAILED = "failed";
+
+export type JobStatus =
+	| typeof JOB_STATUS_PENDING
+	| typeof JOB_STATUS_PROCESSING
+	| typeof JOB_STATUS_COMPLETED
+	| typeof JOB_STATUS_FAILED;
+
+// ===================================================================================
+// Telemetry and Metrics Definitions
+// ===================================================================================
+
+export const GENERATE_CLICK_EVENT = "generate_click";
+export const DOWNLOAD_CLICK_EVENT = "download_click";
+export const IMAGE_GENERATED_EVENT = "image_generated";
+export const METRIC_STATUS_SUCCESS = "success";
+export const METRIC_STATUS_ERROR = "error";
+export const METRIC_STATUS_VALIDATION_ERROR = "validation_error";
+
+export type WcagLevel = "AAA" | "AA" | "FAIL";
+export type MetricStatus =
+	| typeof METRIC_STATUS_SUCCESS
+	| typeof METRIC_STATUS_ERROR
+	| typeof METRIC_STATUS_VALIDATION_ERROR;
+
+export type EventType =
+	| typeof GENERATE_CLICK_EVENT
+	| typeof DOWNLOAD_CLICK_EVENT
+	| typeof IMAGE_GENERATED_EVENT;
+
+export interface MetricPayload {
+	// Core event data
+	event: EventType | string;
+	timestamp: string;
+	status: MetricStatus;
+	errorMessage?: string;
+
+	// Cover generation data
+	size?: {
+		width: number;
+		height: number;
+	};
+	font?: string;
+	titleLength?: number;
+	subtitleLength?: number;
+
+	// Form/Validation data
+	contrastRatio?: number;
+	wcagLevel?: WcagLevel;
+
+	// Performance/Timing data
+	duration?: number; // Backend generation duration (ms)
+	clientDuration?: number; // Client-side duration (ms)
+}
+
+// ===================================================================================
+// Analytics Metrics Schema Definitions
+// ===================================================================================
+
+export type WcagTrendItem = {
+	date: string;
+} & Partial<Record<WcagLevel, number>>;
+
+export type LengthDistribution = {
+	short: number;
+	medium: number;
+	long: number;
+};
+
+export type SubtitleDistribution = {
+	none: number; // 0 chars (subtitle-specific)
+} & LengthDistribution;
+
+export type HourlyTrendItem = {
+	hour: number; // 0-23
+	count: number;
+};
+
+export type WeeklyTrendItem = {
+	week: string; // "2024-W01" format
+	percentage: number;
+};
+
+export type PerformanceBySize = {
+	size: string; // preset label
+	avgBackendDuration: number;
+	p95BackendDuration: number;
+	avgClientDuration: number;
+	p95ClientDuration: number;
+};
+
+export type UserEngagement = {
+	uiGenerationAttempts: number;
+	totalDownloads: number;
+	downloadRate: number;
+	dailyTrend: Array<{ date: string; count: number }>;
+	totalSuccessfulGenerations: number;
+	uiUsagePercent: number;
+	apiUsagePercent: number;
+	hourlyTrend: HourlyTrendItem[];
+};
+
+export type FeaturePopularity = {
+	topFonts: Array<{ font: string; count: number }>;
+	topSizes: Array<{ size: string; count: number }>;
+	titleLengthStats: {
+		_id: null;
+		avgTitleLength: number;
+		minTitleLength: number;
+		maxTitleLength: number;
+	};
+	titleLengthDistribution: LengthDistribution;
+	subtitleUsagePercent: number;
+	subtitleUsageDistribution: SubtitleDistribution;
+	subtitleTrendOverTime: WeeklyTrendItem[];
+};
+
+export type AccessibilityCompliance = {
+	wcagDistribution: Array<{ level: string; count: number }>;
+	contrastStats: {
+		_id: null;
+		avgContrastRatio: number;
+		minContrastRatio: number;
+		maxContrastRatio: number;
+	};
+	wcagTrend: WcagTrendItem[];
+};
+
+export type PerformanceMetrics = {
+	backendPerformance: {
+		avgBackendDuration: number;
+		minBackendDuration: number;
+		maxBackendDuration: number;
+		p50BackendDuration: number;
+		p95BackendDuration: number;
+		p99BackendDuration: number;
+		backendDurationTrend: Array<{ date: string; avgDuration: number }>;
+	};
+	clientPerformance: {
+		avgClientDuration: number;
+		minClientDuration: number;
+		maxClientDuration: number;
+		p50ClientDuration: number;
+		p95ClientDuration: number;
+		p99ClientDuration: number;
+		clientDurationTrend: Array<{ date: string; avgDuration: number }>;
+	};
+	networkLatency: {
+		avgNetworkLatency: number;
+	};
+	performanceBySize: PerformanceBySize[];
+};
+
+export type AnalyticsResult = {
+	userEngagement: UserEngagement;
+	featurePopularity: FeaturePopularity;
+	accessibilityCompliance: AccessibilityCompliance;
+	performanceMetrics: PerformanceMetrics;
+};
+
+// ===================================================================================
+// Validation Constants
+// ===================================================================================
+
+// Title and Subtitle length limits
+export const MAX_TITLE_LENGTH = 40;
+export const MAX_SUBTITLE_LENGTH = 70;
+
+// Title length distribution percentages for analytics
+export const LENGTH_DISTRIBUTION = {
+	SHORT_PERCENT: 0.25, // 25% of MAX_TITLE_LENGTH
+	MEDIUM_PERCENT: 0.625, // 62.5% of MAX_TITLE_LENGTH
+	LONG_PERCENT: 1.25, // 125% of MAX_TITLE_LENGTH (can exceed max)
+} as const;
+
+export const MAX_BATCH_SIZE = 5;
+
+export const TITLE_LENGTH_THRESHOLDS = {
+	SHORT_MAX: Math.floor(MAX_TITLE_LENGTH * LENGTH_DISTRIBUTION.SHORT_PERCENT),
+	MEDIUM_MAX: Math.floor(MAX_TITLE_LENGTH * LENGTH_DISTRIBUTION.MEDIUM_PERCENT),
+	LONG_MAX: Math.floor(MAX_TITLE_LENGTH * LENGTH_DISTRIBUTION.LONG_PERCENT),
+} as const;
+
+export const SUBTITLE_LENGTH_THRESHOLDS = {
+	SHORT_MAX: Math.floor(
+		MAX_SUBTITLE_LENGTH * LENGTH_DISTRIBUTION.SHORT_PERCENT,
+	),
+	MEDIUM_MAX: Math.floor(
+		MAX_SUBTITLE_LENGTH * LENGTH_DISTRIBUTION.MEDIUM_PERCENT,
+	),
+	LONG_MAX: Math.floor(MAX_SUBTITLE_LENGTH * LENGTH_DISTRIBUTION.LONG_PERCENT),
+} as const;
+
+export const SIZE_RANGE = { min: 1, max: 1200 };
+export const SIZE_PRESETS = [
+	{ label: "Post (1200 × 627)", width: 1200, height: 627 },
+	{ label: "Square (1080 × 1080)", width: 1080, height: 1080 },
+];
+
+export const FONT_OPTIONS = [
+	"Montserrat",
+	"Roboto",
+	"Lato",
+	"Playfair Display",
+	"Open Sans",
+] as const;
+
+export const WCAG_AA_THRESHOLD = 4.5;
+export const HEX_COLOR_REGEX = /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/;
+export const DEFAULT_FILENAME = "cover";
+
+// ===================================================================================
+// Derived Types
+// ===================================================================================
+
+export type AllowedFont = (typeof FONT_OPTIONS)[number];
+
+export interface ImageParams {
+	width: number;
+	height: number;
+	backgroundColor: string;
+	textColor: string;
+	font: string;
+	title: string;
+	subtitle?: string;
+	filename: string;
+}
+
+export interface ValidationError {
+	field: string;
+	message: string;
+}
+
+// ===================================================================================
+// Validation Functions (Color & Contrast)
+// ===================================================================================
+
+export function hexToRgb(
+	hex: string,
+): { r: number; g: number; b: number } | null {
+	if (!hex || typeof hex !== "string" || !HEX_COLOR_REGEX.test(hex)) {
+		return null;
+	}
+
+	const cleaned = hex.replace("#", "");
+	const hexValue =
+		cleaned.length === 3
+			? cleaned
+					.split("")
+					.map((c) => c + c)
+					.join("")
+			: cleaned;
+
+	const num = parseInt(hexValue, 16);
+	return {
+		r: (num >> 16) & 255,
+		g: (num >> 8) & 255,
+		b: num & 255,
+	};
+}
+
+export function getRelativeLuminance(rgb: {
+	r: number;
+	g: number;
+	b: number;
+}): number {
+	const [r, g, b] = [rgb.r, rgb.g, rgb.b].map((c) => {
+		const normalized = c / 255;
+		return normalized <= 0.03928
+			? normalized / 12.92
+			: ((normalized + 0.055) / 1.055) ** 2.4;
+	});
+
+	return 0.2126 * r + 0.7152 * g + 0.0722 * b;
+}
+
+export function getContrastRatio(
+	color1: string,
+	color2: string,
+): number | null {
+	const rgb1 = hexToRgb(color1);
+	const rgb2 = hexToRgb(color2);
+
+	if (!rgb1 || !rgb2) {
+		return null;
+	}
+
+	const lum1 = getRelativeLuminance(rgb1);
+	const lum2 = getRelativeLuminance(rgb2);
+
+	const lighter = Math.max(lum1, lum2);
+	const darker = Math.min(lum1, lum2);
+
+	return (lighter + 0.05) / (darker + 0.05);
+}
+
+export function getWCAGLevelFromRatio(ratio: number): "AAA" | "AA" | "FAIL" {
+	if (ratio >= 7) return "AAA";
+	if (ratio >= WCAG_AA_THRESHOLD) return "AA";
+	return "FAIL";
+}
+
+export function meetsWCAGAA(ratio: number): boolean {
+	return ratio >= WCAG_AA_THRESHOLD;
+}
+
+export function getContrastStatus(
+	backgroundColor: string,
+	textColor: string,
+): {
+	status: "good" | "warning" | "poor";
+	ratio: number | null;
+	level: "AAA" | "AA" | "FAIL" | null;
+	message: string;
+} {
+	const ratio = getContrastRatio(backgroundColor, textColor);
+
+	if (ratio === null) {
+		return {
+			status: "poor",
+			ratio: null,
+			level: null,
+			message: "Invalid color format",
+		};
+	}
+
+	const level = getWCAGLevelFromRatio(ratio);
+	const status =
+		level === "FAIL" ? "poor" : level === "AA" ? "warning" : "good";
+
+	const messages: Record<string, string> = {
+		AAA: `${ratio.toFixed(2)}:1 ✓ AAA (Enhanced)`,
+		AA: `${ratio.toFixed(2)}:1 ✓ AA (Standard)`,
+		FAIL: `${ratio.toFixed(2)}:1 ✗ Below AA (Poor)`,
+	};
+
+	return {
+		status,
+		ratio,
+		level,
+		message: messages[level],
+	};
+}
+
+// ===================================================================================
+// Validation Messages and Field Validation Functions
+// ===================================================================================
+
+export const VALIDATION_MESSAGES = {
+	TITLE_REQUIRED: "Title is required",
+	TITLE_TOO_LONG: `Title must be ${MAX_TITLE_LENGTH} characters or less`,
+	SUBTITLE_TOO_LONG: `Subtitle must be ${MAX_SUBTITLE_LENGTH} characters or less`,
+	INVALID_WIDTH: `Width must be between ${SIZE_RANGE.min} and ${SIZE_RANGE.max}`,
+	INVALID_HEIGHT: `Height must be between ${SIZE_RANGE.min} and ${SIZE_RANGE.max}`,
+	INVALID_COLOR: "Color must be a valid HEX format (e.g., #FFFFFF or #FFF)",
+	INVALID_FONT: `Font must be one of: ${FONT_OPTIONS.join(", ")}`,
+	INVALID_CONTRAST: `Contrast ratio must be at least ${WCAG_AA_THRESHOLD} for WCAG AA compliance`,
+	BATCH_EMPTY: "Batch request must contain at least one item",
+	BATCH_TOO_LARGE: `Batch request cannot exceed ${MAX_BATCH_SIZE} items`,
+} as const;
+
+export function validateSize(
+	width: number | undefined,
+	height: number | undefined,
+): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	if (
+		width === undefined ||
+		!Number.isInteger(width) ||
+		Number.isNaN(width) ||
+		width < SIZE_RANGE.min ||
+		width > SIZE_RANGE.max
+	) {
+		errors.push({
+			field: "width",
+			message: VALIDATION_MESSAGES.INVALID_WIDTH,
+		});
+	}
+
+	if (
+		height === undefined ||
+		!Number.isInteger(height) ||
+		Number.isNaN(height) ||
+		height < SIZE_RANGE.min ||
+		height > SIZE_RANGE.max
+	) {
+		errors.push({
+			field: "height",
+			message: VALIDATION_MESSAGES.INVALID_HEIGHT,
+		});
+	}
+
+	return errors;
+}
+
+export function validateColors(
+	backgroundColor: string,
+	textColor: string,
+): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	if (!HEX_COLOR_REGEX.test(backgroundColor)) {
+		errors.push({
+			field: "backgroundColor",
+			message: VALIDATION_MESSAGES.INVALID_COLOR,
+		});
+	}
+
+	if (!HEX_COLOR_REGEX.test(textColor)) {
+		errors.push({
+			field: "textColor",
+			message: VALIDATION_MESSAGES.INVALID_COLOR,
+		});
+	}
+
+	return errors;
+}
+
+export function validateFont(font: string): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	if (!FONT_OPTIONS.includes(font as AllowedFont)) {
+		errors.push({
+			field: "font",
+			message: VALIDATION_MESSAGES.INVALID_FONT,
+		});
+	}
+
+	return errors;
+}
+
+export function validateTextLength(
+	title: string,
+	subtitle?: string,
+): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	if (title && title.length > MAX_TITLE_LENGTH) {
+		errors.push({
+			field: "title",
+			message: VALIDATION_MESSAGES.TITLE_TOO_LONG,
+		});
+	}
+
+	if (subtitle && subtitle.length > MAX_SUBTITLE_LENGTH) {
+		errors.push({
+			field: "subtitle",
+			message: VALIDATION_MESSAGES.SUBTITLE_TOO_LONG,
+		});
+	}
+
+	return errors;
+}
+
+export function generateFilename(
+	filename?: string,
+	defaultName: string = DEFAULT_FILENAME,
+): string {
+	if (!filename || filename.trim().length === 0) {
+		const timestamp = Math.floor(Date.now() / 1000);
+		return `${defaultName}-${timestamp}.png`;
+	}
+	return filename;
+}
+
+export function validateContrast(
+	backgroundColor: string,
+	textColor: string,
+): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	const ratio = getContrastRatio(backgroundColor, textColor);
+
+	if (ratio === null || ratio < WCAG_AA_THRESHOLD) {
+		errors.push({
+			field: "contrast",
+			message: VALIDATION_MESSAGES.INVALID_CONTRAST,
+		});
+	}
+
+	return errors;
+}
+
+export function validateImageParams(params: ImageParams): ValidationError[] {
+	const errors: ValidationError[] = [];
+
+	errors.push(...validateSize(params.width, params.height));
+	errors.push(...validateColors(params.backgroundColor, params.textColor));
+	errors.push(...validateFont(params.font));
+	errors.push(...validateTextLength(params.title, params.subtitle));
+	errors.push(...validateContrast(params.backgroundColor, params.textColor));
+
+	return errors;
+}
+
+export function validateBatchRequest(
+	requests: ImageParams[],
+): ValidationError[] {
+	if (!requests || requests.length === 0) {
+		return [{ field: "batch", message: VALIDATION_MESSAGES.BATCH_EMPTY }];
+	}
+
+	if (requests.length > MAX_BATCH_SIZE) {
+		return [{ field: "batch", message: VALIDATION_MESSAGES.BATCH_TOO_LARGE }];
+	}
+
+	const errors: ValidationError[] = [];
+	requests.forEach((req, index) => {
+		const itemErrors = validateImageParams(req);
+		if (itemErrors.length > 0) {
+			errors.push(
+				...itemErrors.map((e) => ({
+					field: `requests[${index}].${e.field}`,
+					message: e.message,
+				})),
+			);
+		}
+	});
+
+	return errors;
+}
